clean up plugglable discovery and refactor serial-discovery usage
