+++ commands/board/list.go
@@ -18,44 +18,50 @@
 package board
 
 import (
-	"context"
-	"errors"
-	"fmt"
-
 	"github.com/arduino/arduino-cli/commands"
 	rpc "github.com/arduino/arduino-cli/rpc/commands"
+	"github.com/pkg/errors"
 )
 
 // List FIXMEDOC
-func List(ctx context.Context, req *rpc.BoardListReq) (*rpc.BoardListResp, error) {
-	pm := commands.GetPackageManager(req)
+func List(instanceID int32) (*rpc.BoardListResp, error) {
+	pm := commands.GetPackageManager(instanceID)
 	if pm == nil {
 		return nil, errors.New("invalid instance")
 	}
 
+	serialDiscovery, err := commands.NewBuiltinSerialDiscovery(pm)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to instance serial-discovery")
+	}
+
+	if err := serialDiscovery.Start(); err != nil {
+		return nil, errors.Wrap(err, "unable to start serial-discovery")
+	}
+	defer serialDiscovery.Close()
+
 	resp := &rpc.BoardListResp{Ports: []*rpc.DetectedPort{}}
-	for _, disc := range commands.GetDiscoveries(req) {
-		ports, err := disc.List()
-		if err != nil {
-			fmt.Printf("Error getting port list from discovery %s: %s\n", disc.ID, err)
-			continue
+
+	ports, err := serialDiscovery.List()
+	if err != nil {
+		return nil, errors.Wrap(err, "error getting port list from serial-discovery")
+	}
+
+	for _, port := range ports {
+		b := []*rpc.BoardListItem{}
+		for _, board := range pm.IdentifyBoard(port.IdentificationPrefs) {
+			b = append(b, &rpc.BoardListItem{
+				Name: board.Name(),
+				FQBN: board.FQBN(),
+			})
 		}
-		for _, port := range ports {
-			b := []*rpc.BoardListItem{}
-			for _, board := range pm.IdentifyBoard(port.IdentificationPrefs) {
-				b = append(b, &rpc.BoardListItem{
-					Name: board.Name(),
-					FQBN: board.FQBN(),
-				})
-			}
-			p := &rpc.DetectedPort{
-				Address:       port.Address,
-				Protocol:      port.Protocol,
-				ProtocolLabel: port.ProtocolLabel,
-				Boards:        b,
-			}
-			resp.Ports = append(resp.Ports, p)
+		p := &rpc.DetectedPort{
+			Address:       port.Address,
+			Protocol:      port.Protocol,
+			ProtocolLabel: port.ProtocolLabel,
+			Boards:        b,
 		}
+		resp.Ports = append(resp.Ports, p)
 	}
 
 	return resp, nil
